Remove src/meson directory
It does not contain much, better merge what's remaining into toplevel
src/ dir. The split is a bit arbitrary anyway, since the whole extension
is about meson.
